+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 otj-kafka changelog
 ===================
 
+3.2.1
+-----
+* Move LogRestoreListener to new module `otj-kafka-streams`. If you are using this, you'll need
+to pull in the dependency declaration
+
+```$xslt
+<dependency>
+    <groupId>com.opentable.components</groupId>
+    <artifactId>otj-kafka-streams</artifactId>
+</dependency>
+```
+
+* Add experimental support for Kafka Streams via builder. This is a work in progress.
+* Rearrange logging logic to not overlog, and add to otl the opentracing headers.
+* Add `otj-kafka-mirrormaker` module with a utility for dynamically changing topic names when mirroring them.
+
 3.2.0
 -----
 * Add spring factory bean KafkaBuilderFactoryBean
@@ -17,7 +33,7 @@ and forward to graphite.
 
 Detailed instructions for usage are in the README.MD.
 
-* Move EmbeddedKafka and Embedded Zookeeper to otj-kafka-tests module. If you use these you may
+* Move EmbeddedKafka and Embedded Zookeeper to `otj-kafka-tests` module. If you use these you may
 still access them via the following dependency declaration
 
 ```$xslt